Everything runs from main.py (required renaming)
