+++ .gitignore
@@ -1,3 +1,2 @@
-*.mlxx
-TOC.html
+files/
 __pycache__
