--- downloadlxx.bat
@@ -1,2 +1,0 @@
-mkdir files
-python main.py
